pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmx-market-maker"
version = "0.1.0"
description = "GMX v2 做市策略机器人"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "fangchen"}
]
keywords = ["gmx", "defi", "market-making", "arbitrum"]

dependencies = [
    "web3>=6.10.0,<7.0.0",
    "numpy>=1.24.0",
    "pandas>=1.4.2",
    "pyyaml>=6.0",
    "hexbytes>=0.3.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
telegram = [
    "python-telegram-bot>=20.0",
]

[project.scripts]
gmx-mm = "gmx_mm.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
